fix: kb/mu4e-main-bury-buffer occasionally failing
It used to fail when opening an e.g. mu4e-headers-mode buffer (e.g. because of
an org link). Since I would not have invoked an mu4e-main buffer, when trying to
quit out of mu4e, no window configuration was stored to restore to. In such
cases, I choose to bury the buffer instead.